Replace non-C CMS with C
